Fix an initialization bug
mLastScroll is used to compute relative scroll offsets based on the last position that we scrolled to. However, if a new time is set in the ScrollLayout early enough in the initialization process, it would compute a relative scroll based on an uninitialized value of mLastScroll (0), corrupting the state of the ScrollLayout. It would correct itself as soon as the user scrolled it a bit, but this sets it to the correct value up front so this bug doesn't happen anymore.